requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-yasg
gunicorn
psycopg2-binary
pymongo
dnspython>=2.4
djongo
elasticsearch>=8.11.1
django-elasticsearch-dsl>=7.3
celery
redis
kombu
amqp
django-storages[boto3]
boto3
ruff
ruff[format]
pytest
pytest-django
pytest-asyncio
coverage
pytest-cov
python-decouple
drf-spectacular
whitenoise
ipython
django-extensions
black
django-redis
django-celery-beat
requests
beautifulsoup4
lxml
selenium
beautifulsoup4
